# webapp/src/main/webapp/docs/swagger.yaml
openapi: 3.0.1
info:
  title: HealthSync API
  version: 1.0.0
...
paths:
...
  /doctors:
    get:
      operationId: listDoctors
      description: Returns a filtered page of doctors
      tags:
        - doctors
      parameters:
        - name: page
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 10
        - name: name
          in: query
          schema:
            type: string
            example: John Doe
        - name: date
          in: query
          schema:
            format: date
            type: string
            example: 2024-07-25
        - name: fromTime
          in: query
          schema:
            allOf:
              - $ref : '#/components/schemas/TimeBlock'
        - name: toTime
          in: query
          schema:
            allOf:
              - $ref : '#/components/schemas/TimeBlock'
        - name: specialty
          in: query
          schema:
            uniqueItems: true
            type: array
            items:
              allOf:
                - $ref : '#/components/schemas/Specialty'
        - name: city
          in: query
          schema:
            uniqueItems: true
            type: array
            items:
              type: string
        - name: healthInsurance
          in: query
          schema:
            uniqueItems: true
            type: array
            items:
              allOf:
                - $ref : '#/components/schemas/HealthInsurance'
        - name: minRating
          in: query
          schema:
            minimum: 1
            maximum: 5
            type: integer
            format: int32
      responses:
        '200':
          description: Returned page of doctors
          content:
            application/vnd.doctor-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref : '#/components/schemas/Doctor'
        '204':
          description: No content for specified page and filters
    post:
      operationId: createDoctor
      description: Creates a doctor
      tags:
        - doctors
      requestBody:
        content:
          application/vnd.doctor.v1+json:
            schema:
              $ref: '#/components/schemas/DoctorRegisterForm'
      responses:
        '201':
          description: Doctor Created
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email already in use
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
...
  /doctors/{doctorId}/vacations:
    get:
      operationId: listVacations
      description: Paginated list of vacations
      tags:
        - vacations
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: Page of vacations
          content:
            application/vnd.vacation-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Vacation'
        '204':
          description: No content for page
        '403':
          description: Only the doctor can create its vacation
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Doctor not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
    post:
      operationId: createVacation
      description: Creates a vacation
      tags:
        - vacations
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/vnd.vacation-form.v1+json:
            schema:
              $ref: '#/components/schemas/DoctorVacationForm'
      responses:
        '201':
          description: Created the vacation
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '403':
          description: Only the doctor can create its vacation
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Doctor not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '409':
          description: There already was a vaction in the specified range
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
  /doctors/{doctorId}/vacations/{vacationId}:
    get:
      operationId: getVacation
      description: Get a vacation
      tags:
        - vacations
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: vacationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returns the vacation
          content:
            application/vnd.vacation.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Vacation'
        '403':
          description: Only the doctor can see its vacation
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Either doctor or vacation does not exist
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteVacation
      description: Delete a vacation
      tags:
        - vacations
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: vacationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Ok
        '403':
          description: Only the doctor can delete its vacation
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
components:
  schemas:
...
    DoctorRegisterForm:
      required:
        - address
        - city
        - confirmPassword
        - email
        - healthInsurances
        - lastname
        - name
        - password
        - specialty
      type: object
      properties:
        password:
          maxLength: 50
          minLength: 4
          pattern: '[a-zA-Z0-9]+'
          type: string
          example: 'johndoesecret1234'
        confirmPassword:
          maxLength: 50
          minLength: 4
          pattern: '[a-zA-Z0-9]+'
          type: string
          example: 'johndoesecret1234'
        name:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'john'
        lastname:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'doe'
        email:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z0-9.+-ñÑ]+@[a-zA-Z0-9.-]+(.com|.com.ar|.edu.ar)'
          type: string
          example: 'john@gmail.con'
        healthInsurance:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/HealthInsurance'
          example: ['OSDE', 'OMINT']
        city:
          maxLength: 20
          minLength: 1
          pattern: '[a-zA-Z0-9. ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'CABA'
        address:
          maxLength: 100
          minLength: 1
          pattern: '[a-zA-Z0-9. ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'San Martin 202'
        specialty:
          allOf:
            - $ref: '#/components/schemas/Specialty'
...
    DoctorVacationForm:
      required:
        - fromDate
        - fromTime
        - toDate
        - toTime
      type: object
      properties:
        fromDate:
          type: string
          format: date
          example: '2024-07-23'
        fromTime:
          allOf:
            - $ref: '#/components/schemas/TimeBlock'
          example: '14:30'
        toDate:
          type: string
          format: date
          example: '2024-07-26'
        toTime:
          allOf:
            - $ref: '#/components/schemas/TimeBlock'
          example: '08:00'
        cancelAppointments:
          type: boolean
        cancelReason:
          type: string
          example: "Me voy de vacaciones"
    Error:
      type: object
      properties:
        message:
          type: string
    ValidationError:
      type: object
      properties:
        attribute:
          type: string
        invalidValue:
          type: string
        message:
          type: string
    Vacation:
      type: object
      properties:
        doctor:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20"
        fromDate:
          type: string
          format: date
          example: "2026-01-01"
        fromTime:
          type: string
          example: "08:00"
        self:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/vacations/7"
        toDate:
          type: string
          format: date
          example: "2027-01-01"
        toTime:
          type: string
          example: "08:30"
...
    Doctor:
      type: object
      properties:
        address:
          type: string
          example: "San Martin 202"
        appointments:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/appointments?userId=20"
        attendingHours:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/attendinghours"
        occupiedHours:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/occupiedhours"
        city:
          type: string
          example: "CABA"
        email:
          type: string
          format: email
          example: "john@gmail.com"
        firstName:
          type: string
          example: "john"
        healthInsurances:
          type: array
          items:
            type: string
            format: uri
            example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/3"
          example:
            - "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/3"
            - "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/2"
            - "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/1"
            - "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/0"
        lastName:
          type: string
          example: "doe"
        notifications:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/notifications?userId=20"
        rating:
          type: number
          format: float
          example: 3.625
        ratingCount:
          type: integer
          format: int32
          example: 8
        reviews:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/reviews"
        self:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20"
        specialty:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/specialties/17"
        vacations:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/vacations"
...
    TimeBlock:
      type: string
      enum:
        - "00:00"
        - "00:30"
        - "01:00"
        - "01:30"
        - "02:00"
        - "02:30"
        - "03:00"
        - "03:30"
        - "04:00"
        - "04:30"
        - "05:00"
        - "05:30"
        - "06:00"
        - "06:30"
        - "07:00"
        - "07:30"
        - "08:00"
        - "08:30"
        - "09:00"
        - "09:30"
        - "10:00"
        - "10:30"
        - "11:00"
        - "11:30"
        - "12:00"
        - "12:30"
        - "13:00"
        - "13:30"
        - "14:00"
        - "14:30"
        - "15:00"
        - "15:30"
        - "16:00"
        - "16:30"
        - "17:00"
        - "17:30"
        - "18:00"
        - "18:30"
        - "19:00"
        - "19:30"
        - "20:00"
        - "20:30"
        - "21:00"
        - "21:30"
        - "22:00"
        - "22:30"
        - "23:00"
        - "23:30"
    HealthInsurance:
      type: string
      enum:
        - NONE
        - OMINT
        - OSDE
        - SWISS_MEDICAL
        - GALENO
        - MEDIFE
        - MEDICUS
        - ACCORD
        - HOSPITAL_ITALIANO
        - HOSPITAL_ALEMAN
        - HOSPITAL_BRITANICO
        - PREVENCION_SALUD
        - AVALIAN
        - WILLIAM_HOPE
        - PRE_MEDIC
        - OSFE
        - OSPACA
        - ANDAR
        - LUIS_PASTEUR
        - UNION_PERSONAL
        - JERARQUICOS
        - OSDEPYM
        - OSPE
        - OSPLAD
        - OSSEG
        - SANCOR_SALUD
    Specialty:
      type: string
      enum:
        - ALLERGY_AND_IMMUNOLOGY
        - PEDIATRIC_ALLERGY_AND_IMMUNOLOGY
        - PATHOLOGICAL_ANATOMY
        - ANESTHESIOLOGY
        - GENERAL_ANGIOLOGY_AND_HEMODYNAMICS
        - CARDIOLOGY
        - CHILDRENS_CARDIOLOGIST
        - CARDIOVASCULAR_SURGERY
        - PEDIATRIC_CARDIOVASCULAR_SURGERY
        - HEAD_AND_NECK_SURGERY
        - THORAX_SURGERY
        - GENERAL_SURGERY
        - CHILD_SURGERY
        - PLASTIC_AND_RECONSTRUCTIVE_SURGERY
        - PERIPHERAL_VASCULAR_SURGERY
        - MEDICAL_CLINIC
        - COLOPROCTOLOGY
        - DERMATOLOGY
        - PEDIATRIC_DERMATOLOGY
        - IMAGE_DIAGNOSIS
        - CARDIAC_ELECTROPHYSIOLOGY
        - EMERGENTOLOGY
        - ENDOCRINOLOGY
        - INFANT_ENDOCRINOLOGIST
        - CLINICAL_PHARMACOLOGY
        - PHYSIATRY
        - GASTROENTEROLOGY
        - CHILDRENS_GASTROENTEROLOGIST
        - MEDICAL_GENETICS
        - GERIATRICS
        - GYNECOLOGY
        - PEDIATRIC_HEMATO_AWARENESS
        - HEMATOLOGY
        - HEMOTHERAPY_AND_IMMUNOHEMATOLOGY
        - HEPATOLOGY
        - PEDIATRIC_HEPATOLOGY
        - INFECTOLOGY
        - CHILD_INFECTOLOGIST
        - SPORTS_MEDICINE
        - WORK_MEDICINE
        - GENERAL_MEDICINE_AND_OR_FAMILY_MEDICINE
        - LEGAL_MEDICINE
        - NUCLEAR_MEDICINE
        - PALITATIVE_MEDICINE
        - NEPHROLOGY
        - CHILD_NEPHROLOGIST
        - NEONATOLOGY
        - PNEUMONOLOGY
        - CHILD_PNEUMONOLOGIST
        - NEUROSURGERY
        - NEUROLOGY
        - CHILD_NEUROLOGIST
        - NUTRITION
        - OBSTETRICS
        - OPHTHALMOLOGY
        - ONCOLOGY
        - ORTHOPEDICS_AND_TRAUMATOLOGY
        - ORTHOPEDICS_AND_CHILD_TRAUMATOLOGY
        - OTORHINOLARYNGOLOGY
        - PEDIATRICS
        - PSYCHIATRY
        - CHILD_AND_YOUTH_PSYCHIATRY
        - RADIOTHERAPY_OR_RADIATION_THERAPY
        - RHEUMATOLOGY
        - CHILD_RHEUMATOLOGIST
        - INTENSIVE_THERAPY
        - INTENSIVE_CHILDRENS_THERAPIST
        - TOCOGINECOLOGY
        - TOXICOLOGY
        - UROLOGY